+++ tests/test_utils.py
@@ -1,31 +1,34 @@
+import logging
 import pytest
-from charset_normalizer.utils import is_accentuated, cp_similarity
+from charset_normalizer.utils import is_accentuated, cp_similarity, set_logging_handler
 
 
 @pytest.mark.parametrize(
     "character, expected_is_accentuated",
     [
-        ('é', True),
-        ('è', True),
-        ('à', True),
-        ('À', True),
-        ('Ù', True),
-        ('ç', True),
-        ('a', False),
-        ('€', False),
-        ('&', False),
-        ('Ö', True),
+        ("é", True),
+        ("è", True),
+        ("à", True),
+        ("À", True),
+        ("Ù", True),
+        ("ç", True),
+        ("a", False),
+        ("€", False),
+        ("&", False),
+        ("Ö", True),
         ("ü", True),
         ("ê", True),
-        ('Ñ', True),
-        ('Ý', True),
-        ('Ω', False),
-        ('ø', False),
-        ('Ё', False),
-    ]
+        ("Ñ", True),
+        ("Ý", True),
+        ("Ω", False),
+        ("ø", False),
+        ("Ё", False),
+    ],
 )
 def test_is_accentuated(character, expected_is_accentuated):
-    assert is_accentuated(character) is expected_is_accentuated, "is_accentuated behavior incomplete"
+    assert (
+        is_accentuated(character) is expected_is_accentuated
+    ), "is_accentuated behavior incomplete"
 
 
 @pytest.mark.parametrize(
@@ -37,7 +40,7 @@ def test_is_accentuated(character, expected_is_accentuated):
         ("latin_1", "iso8859_4", True),
         ("latin_1", "cp1251", False),
         ("cp1251", "mac_turkish", False),
-    ]
+    ],
 )
 def test_cp_similarity(cp_name_a, cp_name_b, expected_is_similar):
     is_similar = cp_similarity(cp_name_a, cp_name_b) >= 0.8